+++ Parse/Parse/Internal/ACL/DefaultACLController/PFDefaultACLController.m
@@ -9,7 +9,11 @@
 
 #import "PFDefaultACLController.h"
 
+#if __has_include(<Bolts/BFTask.h>)
 #import <Bolts/BFTask.h>
+#else
+#import "BFTask.h"
+#endif
 
 #import "PFACLPrivate.h"
 #import "PFAsyncTaskQueue.h"