+++ Parse/Parse/Internal/BFTask+Private.m
@@ -9,8 +9,13 @@
 
 #import "BFTask+Private.h"
 
+#if __has_include(<Bolts/BFExecutor.h>)
 #import <Bolts/BFExecutor.h>
 #import <Bolts/BFTaskCompletionSource.h>
+#else
+#import "BFExecutor.h"
+#import "BFTaskCompletionSource.h"
+#endif
 
 #import "PFLogging.h"