+++ Parse/Parse/Internal/Commands/CommandRunner/URLSession/PFURLSessionCommandRunner.m
@@ -10,7 +10,11 @@
 #import "PFURLSessionCommandRunner.h"
 #import "PFURLSessionCommandRunner_Private.h"
 
+#if __has_include(<Bolts/BFTaskCompletionSource.h>)
 #import <Bolts/BFTaskCompletionSource.h>
+#else
+#import "BFTaskCompletionSource.h"
+#endif
 
 #import "BFTask+Private.h"
 #import "PFAssert.h"