+++ Parse/Parse/Internal/LocalDataStore/OfflineQueryLogic/PFOfflineQueryLogic.m
@@ -9,8 +9,13 @@
 
 #import "PFOfflineQueryLogic.h"
 
+#if __has_include(<Bolts/BFTask.h>)
 #import <Bolts/BFTask.h>
 #import <Bolts/BFExecutor.h>
+#else
+#import "BFTask.h"
+#import "BFExecutor.h"
+#endif
 
 #import "PFACL.h"
 #import "PFAssert.h"