+++ Parse/Parse/Internal/Object/CurrentController/PFCurrentObjectControlling.h
@@ -9,7 +9,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import <Parse/PFConstants.h>
+#import "PFConstants.h"
 
 NS_ASSUME_NONNULL_BEGIN