+++ Parse/Parse/Internal/PFCommandCache.m
@@ -12,8 +12,13 @@
 #include <mach/mach_time.h>
 #include <sys/xattr.h>
 
+#if __has_include(<Bolts/BFTask.h>)
 #import <Bolts/BFTask.h>
 #import <Bolts/BFTaskCompletionSource.h>
+#else
+#import "BFTask.h"
+#import "BFTaskCompletionSource.h"
+#endif
 
 #import "BFTask+Private.h"
 #import "PFAssert.h"