+++ Parse/Parse/Internal/Session/PFSession_Private.h
@@ -9,7 +9,7 @@
 
 #import <Foundation/Foundation.h>
 
-#import <Parse/PFSession.h>
+#import "PFSession.h"
 
 @class PFSessionController;