+++ ParseTwitterUtils/ParseTwitterUtils/Internal/PFTwitterAuthenticationProvider.h
@@ -9,7 +9,11 @@
 
 #import <Foundation/Foundation.h>
 
+#if __has_include(<Parse/PFUserAuthenticationDelegate.h>)
 #import <Parse/PFUserAuthenticationDelegate.h>
+#else
+#import "PFUserAuthenticationDelegate.h"
+#endif
 
 @class BFTask<__covariant BFGenericType>;
 @class PF_Twitter;