+++ ParseTwitterUtils/ParseTwitterUtils/Internal/PFTwitterAuthenticationProvider.m
@@ -9,9 +9,17 @@
 
 #import "PFTwitterAuthenticationProvider.h"
 
+#if __has_include(<Bolts/BFTask.h>)
 #import <Bolts/BFTask.h>
+#else
+#import "BFTask.h"
+#endif
 
+#if __has_include(<Parse/PFConstants.h>)
 #import <Parse/PFConstants.h>
+#else
+#import "PFConstants.h"
+#endif
 
 #import "PFTwitterPrivateUtilities.h"
 #import "PF_Twitter.h"