+++ ParseUI/ParseUIDemo/Classes/CustomViewControllers/QueryTableViewController/PaginatedTableViewController.m
@@ -21,10 +21,15 @@
 
 #import "PaginatedTableViewController.h"
 
+#if __has_include(<Parse/PFObject.h>)
 #import <Parse/PFObject.h>
 #import <Parse/PFQuery.h>
+#else
+#import "PFObject.h"
+#import "PFQuery.h"
+#endif
 
-#import <ParseUI/PFTableViewCell.h>
+#import "PFTableViewCell.h"
 
 @implementation PaginatedTableViewController