+++ ParseUI/SignInWithAppleTests/SignInWithAppleTests.m
@@ -9,7 +9,6 @@
 #import <XCTest/XCTest.h>
 #import "PFAppleUtils.h"
 #import "Parse/PFUser.h"
-#import "PFLoginViewController.h"
 
 @import OCMock;